pyproject.toml
[project]
name = "workman"
version = "0.1.0"
description = "CLI tool for managing a workspace of projects"
readme = "README.md"
authors = [
    { name = "Isaac Henry", email = "isaac@symptomwise.com" }
]
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
wm = "workman.cli:cli"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"
